datatable row actions (use trait to handle row actions)
